internal/core/users_api/api/remove_user_test.go
package api

/**
 * Panther is a Cloud-Native SIEM for the Modern Security Team.
 * Copyright (C) 2020 Panther Labs Inc
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

import (
	"testing"

	"github.com/aws/aws-sdk-go/aws"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"

	"github.com/panther-labs/panther/api/lambda/users/models"
	"github.com/panther-labs/panther/internal/core/users_api/cognito"
)

func TestRemoveUser(t *testing.T) {
	userID := aws.String("user-remove")
	otherUserID := aws.String("user-other")

	mockGateway := &cognito.MockUserGateway{}
	userGateway = mockGateway
	mockGateway.On("ListUsers", &models.ListUsersInput{}).Return(
		[]models.User{
			{ID: userID},
			{ID: otherUserID},
		},
		nil,
	)

	mockGateway.On("DeleteUser", userID).Return(nil)

	result, err := API{}.RemoveUser(&models.RemoveUserInput{
		RequesterID: aws.String(systemUserID),
		ID:          userID,
	})
	require.NoError(t, err)
	assert.Equal(t, &models.RemoveUserOutput{ID: userID}, result)
	mockGateway.AssertExpectations(t)
}
